added service tests

+++ SangivSpring/src/main/java/com/qa/sangivspring/service/ClubService.java
@@ -64,9 +64,12 @@ public class ClubService {
 	
 	//delete
 	public boolean delete(Long ID) {
-		this.repo.deleteById(ID);
-		return !this.repo.existsById(ID);
-	}
+        if (!this.repo.existsById(ID)) {
+            throw new ClubNotFoundException();
+        }
+        this.repo.deleteById(ID);
+        return !this.repo.existsById(ID);
+    }
 	
 	
 }
